% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diagnostics}
\alias{diagnostics}
\title{Diagnostics plots for read counts}
\usage{
diagnostics(path, controls)
}
\arguments{
\item{path}{Path to file with *tradis_gene_insert_sites.csv files}

\item{controls}{A string (whole or partial) to identify the controls (e.g. "control", or "time0")}
}
\description{
Generates diagnostic plots using read counts as input. File names are expected to follow \*_1.tradis_gene_insert_sites.csv, \*_2.tradis_gene_insert_sites.csv. Conditions will be named based on input names before the replicate number. All conditions and controls are expected to have two replicates.
}
