% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_fc.R
\name{diagnostics_fc}
\alias{diagnostics_fc}
\title{Diagnostics plots for log fold changes}
\usage{
diagnostics_fc(path)
}
\arguments{
\item{path}{Path to file with *.csv files}
}
\description{
Generates diagnostic plots of output from tradis_comparison.R.
}
