% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_rc.R
\name{diagnostics_rc}
\alias{diagnostics_rc}
\title{Diagnostics plots for read counts}
\usage{
diagnostics_rc(path)
}
\arguments{
\item{path}{Path to file with *tradis_gene_insert_sites.csv files}
}
\description{
Generates diagnostic plots using read counts as input.
}
