% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sig.R
\name{get_sig}
\alias{get_sig}
\title{Extract significant genes}
\usage{
get_sig(x, save_path, sig = 0.05, abslogFC = 0, subset)
}
\arguments{
\item{x}{output from structure_embl()}

\item{save_path}{Path to directory to save output files. If not provided, will save in current working directory.}

\item{sig}{Significant level for cutoff (default 0.05)}

\item{abslogFC}{Cutoff level for the absolute value of the logFC (default 0)}

\item{subset}{Either "gene" or "all". "Gene" will output gene and logFC information only - great for input into EcoCyc. "all" will output locus tags, gene names and function with logFC data. (default= "all)}
}
\description{
Takes the logFC and significance values as outputted by tradis_comparison.R and makes new files. Will output only gene name and logFC (perfect for input into EcoCyc/BioCyc SmartTables)
}
