% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promoter_weblogo.R
\name{promoter_weblogo}
\alias{promoter_weblogo}
\title{Promoter sequence weblogo}
\usage{
sig_promoters(logfcs, promoters, save_path = "~/Desktop/)
}
\arguments{
\item{x}{The output from structure_embl()}

\item{promoters}{A data frame of gene names and their associated upstream sequences. Easily generated using Artemis (Select --> All CDS features. File --> Write --> Upstream bases of selected features --> FASTA format).}

\item{save_path}{Location to save the promoter weblogo}

\item{save_fasta}{TRUE or FALSE. Default is FALSE. If TRUE, will save to same directory as weblogo}
}
\description{
Matches the significant genes with their promoter regions and creates a weblogo of consensus regions. THIS CURRENTLY DOESN'T WORK FOR rRNA.
}
