% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_info.R
\name{sig_info}
\alias{sig_info}
\title{Get info from CSVs}
\usage{
sig_info(path, abslogfc = 1, sig = 0.05)
}
\arguments{
\item{path}{path to folder containing csv files}

\item{abslogfc}{absolute log fold change cutoff for output. Defaults to 0.}

\item{sig}{significance level cutoff for output. Defaults to 0.05.}
}
\description{
Gives the number of significant genes at log fold change thresholds for all conditions in a file path.
}
