% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_csv.R
\name{structure_csv}
\alias{structure_csv}
\title{Data structuring}
\usage{
structure_csv(csvpath = "")
}
\arguments{
\item{csvpath}{The path to your Bio-TraDIS csv output files.}
}
\value{

}
\description{
Takes Bio-TraDIS output .csvs and restructures data for further tradisanalyseR scripts. All .csv files in the folder will be obtained.
}
