% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_embl.R
\name{structure_embl}
\alias{structure_embl}
\title{Data structuring}
\usage{
setwd("/path/to/csv/folder/)

x <- structure_embl()
}
\arguments{
\item{csvpath}{The path to your Bio-TraDIS csv output files.}

\item{emblpath}{The path to the folder containing the corresponding organism .embl file (optional - leave blank if the embl file is in the same directory as your input files).}
}
\value{

}
\description{
Takes Bio-TraDIS output .csvs and corresponding EMBL file and restructures data for further tradisanalyseR scripts. All .csv files and .embl files will be uploaded.
}
\examples{

}
