% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_operons.R
\name{get_operons}
\alias{get_operons}
\title{Plot transcription units}
\usage{
get_operons(logfcs, x, abslogfc = 0, sig = 0.05)
}
\arguments{
\item{logfcs}{Output from structure_csv()}

\item{x}{Tab separated file of BioCyc transcription unit output.}

\item{abslogfc}{LogFC cutoff for output. Defaults to 0 to obtain all significant genes.}

\item{sig}{Significance level cutoff for output. Defaults to 0.05.}
}
\description{
Takes a tab separated file with lists of genes in transcription units (made with BioCyc in mind) and output from structure_csv() to show transcription unit fold changes.
}
