% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promoter_weblogo.R
\name{promoter_weblogo}
\alias{promoter_weblogo}
\title{Promoter sequence weblogo}
\usage{
promoter_weblogo(
  logfcs,
  promoters,
  save_plot = FALSE,
  save_path,
  save_fasta = FALSE
)
}
\arguments{
\item{logfcs}{The output from structure_csv()}

\item{promoters}{A data frame of locus tags and their associated upstream sequences. Can be generated using Artemis (Select --> All CDS features. File --> Write --> Upstream bases of selected features --> FASTA format). However, will need to edit for locus tags instead of gene name. Need the first column to be named locus_tag.}

\item{save_path}{Location to save the promoter weblogo. Default is the current working directory.}

\item{save_fasta}{TRUE or FALSE. If TRUE, will save to same directory as weblogo. (default = FALSE)}
}
\description{
Matches the significant genes with their promoter regions and creates a weblogo of consensus regions. This currently doesn't work for rRNA.
}
