% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_embl.R
\name{structure_embl}
\alias{structure_embl}
\title{Data structuring}
\usage{
structure_embl()
}
\value{

}
\description{
Takes Bio-TraDIS output .csvs and corresponding EMBL file and restructures data for further tradisanalyseR scripts. Working directory must be set to folder containing the correct files. All .csv files and .embl files will be uploaded.
}
\examples{

}
